-Fix: MSVC blabla

+++ src/unknown/emu_07D4.c
@@ -544,7 +544,7 @@ static void Unknown_07D4_034D(bool arg06, bool arg08, bool arg0A)
 
 			/* ENHANCEMENT -- When fading in the game on start, you don't see the fade as it is against the already drawn screen. */
 			if (g_dune2_enhanced) {
-				uint8 oldScreenID = g_global->screenActiveID;
+				uint16 oldScreenID = g_global->screenActiveID;
 
 				GUI_Screen_SetActive(0);
 				GUI_DrawFilledRectangle(g_curWidgetXBase << 3, g_curWidgetYBase, (g_curWidgetXBase + g_curWidgetWidth) << 3, g_curWidgetYBase + g_curWidgetHeight, 0);
